requirements.txt
fastapi==0.109.0
uvicorn==0.15.0
transformers==4.36.2
Pillow==10.2.0
python-multipart==0.0.6
